bug fix: hibernate returns BigDecimal instead of Long for the query to find max GeneratorIds

+++ trade-consumer/src/main/java/io/alw/css/tradeconsumer/cashflow/model/GeneratorIds.java
@@ -1,7 +1,9 @@
 package io.alw.css.tradeconsumer.cashflow.model;
 
+import java.math.BigDecimal;
+
 public record GeneratorIds(
-        Long tradeId,
-        Long matchEventId
+        BigDecimal tradeId,
+        BigDecimal matchEventId
 ) {
 }
